+++ extensions/imessage-2fa/src/messages.ts
@@ -10,6 +10,7 @@ function getBaseQuery() {
   const preferences = getPreferenceValues<Preferences>();
   const lookBackDays = parseInt(preferences?.lookBackDays || "1") || 1;
   const lookBackMinutes = lookBackDays * 24 * 60;
+  const ignoreRead = preferences.ignoreRead ? 0 : 1;
   return `
     select
       message.guid,
@@ -26,6 +27,7 @@ function getBaseQuery() {
       and message.text is not null
       and length(message.text) > 0
       and datetime(message.date / 1000000000 + strftime('%s', '2001-01-01'), 'unixepoch', 'localtime') >= datetime('now', '-${lookBackMinutes} minutes', 'localtime')
+	  and message.is_read = ${ignoreRead}
   `;
 }