+++ extensions/infakt/src/search_clients.tsx
@@ -1,4 +1,4 @@
-import SearchClients from "./components/clients/SearchClients";
+import { SearchClients } from "@/components/clients/SearchClients";
 
 export default function Command() {
   return <SearchClients />;